Fix errors during GET and POST
This commit irons out some of the errors that popped up while calling
GET and POST routes. In models.py, the commit adds an ordering to the
models because it would not display otherwise. In serializers.py, the
commit fixes some incorrect variable names and establishes the foreign
key relationships that ItemOrder has. Finally, this commit displays all
the nested information in the get requests.